requirements.txt
docker-py #docker library for python operations

